requirements.txt
# Methane Shadow Hunter - Dependencies
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
folium>=0.14.0
requests>=2.31.0
pystac-client>=0.7.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0

# Django REST API
django>=5.2
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
django-filter>=24.0
mysqlclient>=2.2.0

# Google Earth Engine
earthengine-api>=1.4.0
